+++ sei-db/db_engine/litt/dbcache/cached_table.go
@@ -203,7 +203,7 @@ func (c *cachedTable) Destroy() error {
 	return c.base.Destroy()
 }
 
-func (c *cachedTable) SetShardingFactor(shardingFactor uint32) error {
+func (c *cachedTable) SetShardingFactor(shardingFactor uint8) error {
 	return c.base.SetShardingFactor(shardingFactor)
 }
