+++ sei-db/db_engine/litt/test/unlock_test.go
@@ -24,7 +24,7 @@ func TestUnlock(t *testing.T) {
 	config, err := litt.DefaultConfig(volumes...)
 	config.Fsync = false // Disable fsync for faster tests
 	config.TargetSegmentFileSize = 100
-	config.ShardingFactor = uint32(len(volumes))
+	config.ShardingFactor = uint8(len(volumes))
 	require.NoError(t, err)
 
 	db, err := littbuilder.NewDB(config)
